Remove prefix from Go 1.21+ function names
This is explained in the code and comments.

In Go 1.21+ the function name produced by runtime.Frame now includes –
for functions created as a closure and called from a package – a prefix
of the code that originally created that function.

This causes real issues with error grouping, particularly when you
consider the standard pattern of building Go HTTP servers with several
middleware functions.